+++ CHANGES.md
@@ -7,6 +7,7 @@
 * `constraints.txt` included
 * Release step removed from CI
 * Vault directory removed
+* Canary to check STAC API search endpoint (https://github.com/fredliporace/cbers-2-stac/issues/83)
 * Better check for errors in bulk ES operations (https://github.com/fredliporace/cbers-2-stac/issues/74)
 * Enabling API GW log at INFO level (https://github.com/fredliporace/cbers-2-stac/issues/72)